% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boots.R
\name{boot_boot}
\alias{boot_boot}
\title{Boot package Bootstrap Function}
\usage{
boot_boot(x, func = means, B = 1001)
}
\arguments{
\item{x}{Numeric Vector of observations}

\item{func}{Statistic to compute the bootstrap estimate of}

\item{B}{The number of replicatates to take}
}
\value{
List of the numeric estimate of the statistic and the 95\%
confidence intervals
}
\description{
\code{boot_boot} computes the bootstrap estimate from a function
}
\details{
\code{boot_r} is a student-written bootstrap function based on the boot package
which computes the bootstrap estimate of a given statistic
}
\examples{
{
x <- rnorm(100, 10) # random normal distribution with mean = 10
boot_boot(x, 1001) # use a bootstrap approximation of the mean
}

}
\references{
DATA 495 Lectures at Northern Michigan University by John Kloke
}
