% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boots.R
\name{boot_r}
\alias{boot_r}
\title{Bootstrap Function}
\usage{
boot_r(x, func, n, B = 1001)
}
\arguments{
\item{x}{Numeric Vector of observations}

\item{func}{Statistic to compute the bootstrap estimate of}

\item{n}{The number of samples in each bootstrap sample}

\item{B}{The number of replicatates to take}
}
\value{
List of the numeric estimate of the statistic and the 95\%
confidence intervals
}
\description{
\code{boot_r} computes the bootstrap estimate from a function
}
\details{
\code{boot_r} is a student-written bootstrap function which computes the bootstrap
estimate of a given statistic
}
\examples{
{
x <- rnorm(100, 10) # random normal distribution with mean = 10
boot_r(x, mean, 10) # use a bootstrap approximation of the mean
}

}
\references{
DATA 495 Lectures at Northern Michigan University by John Kloke
}
