% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_hle1}
\alias{cpp_hle1}
\alias{HLE1}
\title{cpp_hle1}
\usage{
cpp_hle1(vec)
}
\arguments{
\item{vec}{The numeric vector of which find the Hodges-Lehmann location estimator}
}
\value{
Numeric representing the Hodges-Lehmann location estimator
}
\description{
Compute the Hodges-Lehmann estimate of location
}
\details{
Using the function \code{cpp_median} this function computes the Hodges-Lehmann estimator of location.
This function cannot allocate a vector of the correct size if the number of measurements exceeds
40,000 and so, throws an error.
}
\examples{
{
library(datar)
x <- rnorm(100, 10) 
datar::cpp_hle1(x)
}
}
\references{
DATA495 lectures with Prof. Kloke.
}
\seealso{
\code{\link[=outer_hle1]{outer_hle1()}} The base-R implementation of HLE1
}
