% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_hle2}
\alias{cpp_hle2}
\alias{HLE2}
\title{cpp_hle2}
\usage{
cpp_hle2(x, y)
}
\arguments{
\item{x}{Numeric vector of measurements from the first population}

\item{y}{Numeric vector of measurements from the second population}
}
\value{
Numeric representing the Hodges-Lehmann estimator of shift
}
\description{
Compute the Hodges-Lehmann estimate of shift
}
\details{
Using the function \code{cpp_median} this function computes the Hodges-Lehmann estimator of shift.
This function cannot allocate a vector of the correct size if the number of measurements exceeds
40,000 and so, throws an error.
}
\examples{
{
x <- rnorm(100, 10) 
y <- rnorm(100, 5) 
cpp_hle2(x)
}
}
\references{
DATA495 lectures with Prof. Kloke.
}
\seealso{
\code{\link[=outer_hle2]{outer_hle2()}} The base-R implementation of HLE2
}
