% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_median}
\alias{cpp_median}
\alias{Median}
\title{cpp_median}
\usage{
cpp_median(vec)
}
\arguments{
\item{vec}{The numeric vector of which to take the median}
}
\value{
Numeric of median
}
\description{
Find the median of a numeric vector
}
\details{
Using the c++ function \code{nth_element}, this function places the element which,
if the vector were sorted, would be in the middle position and then returns that same element
}
\examples{
{
x <- rnorm(100, 10) 
cpp_median(x)
}
}
\references{
DATA495 lectures with Prof. Kloke.
}
