% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data}
\alias{data}
\title{Load objects from a data repository}
\usage{
data(
  ...,
  list = character(),
  package = NULL,
  lib.loc = NULL,
  verbose = getOption("verbose"),
  envir = .GlobalEnv,
  overwrite = TRUE,
  paths
)
}
\arguments{
\item{...}{See \code{?utils::data}}
}
\description{
A copy of the \code{utils::data} (4.1.0) function, copied into this namespace to use the \code{datar::find.package} function instead of \code{base::find.package}.
}
