% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{data_repository_path}
\alias{data_repository_path}
\title{Get path to the data repository of a project}
\usage{
data_repository_path()
}
\value{
Character string of path to the data repository for a project.
}
\description{
Checks \code{options()$datarepo.path} for the path, otherwise uses the project path and \code{options()$datarepo.subdir} which defaults to \code{data-repository}.
}
