% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datar.R
\docType{package}
\name{datar}
\alias{datar}
\alias{datar-package}
\title{Data Utilities for R}
\description{
Utility functions to make data manipulation easier and more consistent.
}
\author{
Nathan Lucas
}
