% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datar_paths.R
\name{datar_path}
\alias{datar_path}
\title{Get path to the datar directory}
\usage{
datar_path(root = get_root(), subdir = "data/r")
}
\arguments{
\item{root}{Path to the top-level directory (for a project)}

\item{subdir}{Path within \code{root} into which data can be saved}
}
\description{
Get path to the datar directory
}
