% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataR.R
\name{delete_files_from_excel}
\alias{delete_files_from_excel}
\title{Delete files based on information stored in an Excel file created by the
function `dataR::folder_details()`}
\usage{
delete_files_from_excel(file_path, save = 0)
}
\arguments{
\item{file_path}{The path to the Excel file containing file details.}

\item{save}{The value in the delete column that avoid the deletion of the file.}
}
\value{
None
}
\description{
This function reads an Excel file containing details about files in folders,
filters files based on the "delete" column, and deletes the selected files.
The "delete" column is expected to contain binary values (0 or 1) indicating
whether the corresponding file should be deleted.
}
\examples{
\dontrun{
delete_files_from_excel("details.xlsx", save = 0)
}
}
