% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boots.R
\name{diff_means}
\alias{diff_means}
\title{diff_means}
\usage{
diff_means(x, y)
}
\arguments{
\item{x}{Numeric vector of samples}

\item{y}{Numeric vector of samples}
}
\value{
List of containing estimate and confidence intervals
}
\description{
Calculate an estimate of the shift of location parameter using difference of means
}
\examples{
{
x <- rnorm(100, 5)
y <- rnorm(100, 10)
diff_means(x, y)
}

}
