% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataR.R
\name{file_size}
\alias{file_size}
\title{Get the size of individual files}
\usage{
file_size(file_paths, units = "bytes")
}
\arguments{
\item{file_paths}{A vector of file paths.}

\item{units}{The desired units for file sizes ('bytes', 'kb', 'mb', or 'gb').}
}
\value{
A data frame with filename and size columns.
}
\description{
This function takes a vector of file paths and returns a data frame with the
filename and size of each file. File sizes can be reported in different units
(bytes, kilobytes, megabytes, gigabytes).
}
\examples{
\dontrun{
file_size(c("file1.txt", "file2.txt"), units = 'kb')
}

}
