% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{find.package}
\alias{find.package}
\title{Find paths to loaded packages}
\usage{
find.package(...)
}
\value{
Paths to all loaded packages (prepended with the data repository path) or the path to the specified \code{package}.
}
\description{
Prepends \code{data_repository_path()} to the result of the \code{base::find_package} function.
}
