% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataR.R
\name{folder_details}
\alias{folder_details}
\title{Eases the cleaning of folder by creating an excel with  details about
the files in the folders that you want to clean. Once create the user must
change/delete the zeros in the delete column for the files that must be
deleted.}
\usage{
folder_details(folders, file_path = NULL, units = "bytes")
}
\arguments{
\item{folders}{A vector of folder paths.}

\item{file_path}{The path where the Excel file will be saved.}

\item{units}{The desired units for file sizes ('bytes', 'kb', 'mb', or 'gb').}
}
\value{
None
}
\description{
This function takes a vector of folder paths, retrieves details about files
in those folders, and writes the information to an Excel file.
The details include file names, last update timestamps, file sizes,
proportional sizes to the folder where it is stored, and add a column for potential deletion.
}
\examples{
\dontrun{
folder_details(c("/path/to/folder1", "/path/to/folder2"),
                file_path = "details.xlsx", units = 'kb')
}
}
