% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boots.R
\name{gen_dist}
\alias{gen_dist}
\title{gen_dist}
\usage{
gen_dist(ns, mean, dist = c("normal", "logistic", "exp"), scale = 1)
}
\arguments{
\item{ns}{A numeric vector of the sizes of each sample \strong{on a log 10 scale}}

\item{mean}{A numeric representing the desired mean location}

\item{dist}{String with name of the requested distribution. Available values are "logistic", "normal" and "exp"}

\item{scale}{Numeric representing scale parameter. Only used for the logistic distibution}
}
\value{
A list of samples
}
\description{
Generate a list of disributions of varying sizes
}
\details{
This helper function is primarily for testing. However, I found this family of helper functions rather useful
so I chose to export them. 'gen_dist' is especially powerful when running bootstraps using the 'dual_boot' function
as it outputs the samples in the format needed by that function.
}
\examples{
{
# We want samples of size 10, 100 and 1000
ns <- c(1, 2, 3)
gen_dist(ns, mean = 1, dist = "normal")
}

}
