% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cor.R
\name{get_cor}
\alias{get_cor}
\title{Get data with specific correlation}
\usage{
get_cor(n = NULL, rho = NULL)
}
\arguments{
\item{n}{Number of observations}

\item{rho}{Specified Pearson correlation}
}
\value{
A list of n x and y values between 0 and 1
}
\description{
Get data with specific correlation
}
\examples{
get_cor(n=20, rho=0.6)
}
