% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datar_paths.R
\name{get_root}
\alias{get_root}
\title{Get path to the root of the project}
\usage{
get_root()
}
\value{
Character string of path to project
}
\description{
From the current working directory, splits the path using the platform-specific path separator and combines the first 10 directories into a path.
}
