% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{get_saved_object_digest}
\alias{get_saved_object_digest}
\title{Access a  list of object digests}
\usage{
get_saved_object_digest(...)
}
\arguments{
\item{..}{Character string name(s) of object(s) that have been saved to the data-repository in this session.}
}
\description{
Objects that are saved to the repository have their checksums recorded here.
}
