% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Create the datar directory}
\usage{
init(path = datar_path())
}
\arguments{
\item{path}{Path to datar directory in which \code{R} objects will be saved}
}
\description{
Create the datar directory
}
