% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boots.R
\name{means}
\alias{means}
\title{Mean function}
\usage{
means(x)
}
\arguments{
\item{x}{Numeric Vector of observations}
}
\value{
The numeric mean of the given vector
}
\description{
\code{means} computes the mean of a numeric vector
}
\details{
\code{means} exists only to simplify the passing of the mean statistic into the
\code{boot_boot} function.
}
\examples{
{
x <- rnorm(100, 10) # random normal distribution with mean = 10
means(x) # compute the mean
}

}
\references{
DATA 495 Lectures at Northern Michigan University by John Kloke
}
