% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestv.R
\name{nestv}
\alias{nestv}
\title{Nested Vector}
\usage{
nestv(inner, outer, skip = 0)
}
\arguments{
\item{inner}{?}

\item{outer}{?}

\item{skip}{?}
}
\value{
Returns a nested vector.
}
\description{
Nested Vector
}
\examples{
nestv(4, 3, 1)  #  1 2 3 4   6 7 8 9   11 12 13 14
nestv(4, 3, 2)  #  1 2 3 4   7 8 9 10   13 14 15 16
nestv(3, 4, 1)  #  1 2 3   5 6 7   9 10 11   13 14 15
}
\author{
Nathan Lucas
}
