% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hle1.R
\name{outer_hle1}
\alias{outer_hle1}
\title{Hodges-Lehmann Estimate of Location}
\usage{
outer_hle1(x)
}
\arguments{
\item{x}{Numeric Vector containing the sample}
}
\value{
Numeric Estimate of the location parameter
}
\description{
\code{outer_hle1} computes the Hodges-Lehmann Location Estimator of a Sample
}
\details{
This particular implementation of hle1 is computed using the R Core \code{outer}
function. Outer is less performant when compared to the C++ method.
The estimate should be roughly equivalent, though.
}
\examples{
x <- rnorm(100, 10) # Random Normal with Location Parameter of 10
outer_hle1(x)

}
\references{
DATA 495 Lectures at Northern Michigan University by John Kloke
}
\seealso{
\code{\link[=cpp_hle1]{cpp_hle1()}} The faster, C++, implementation of this function
}
