% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readcsv.R
\name{read_csv}
\alias{read_csv}
\title{Data Input from CSV}
\usage{
read_csv(file, factors = NULL, ...)
}
\arguments{
\item{file}{name of the file from which the data are to be read
(e.g., \code{"csv/mydata.csv"}).}

\item{factors}{range of columns that are factors
(e.g., \code{2:4}, \code{c(1, 3:4)}, etc.).}

\item{...}{further arguments passed to \code{utils::read.csv}.}
}
\value{
A \code{data.frame}.
}
\description{
Import data frame from CSV file with header labels.
}
\examples{
\dontrun{
# Columns 1, 3, and 4 are factors
dat <- datar::read_csv(file = "mydata", factors = c(1, 3:4))
}
}
\author{
Nathan Lucas
}
