% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_datar.R
\name{save_datar}
\alias{save_datar}
\title{Save objects to datar}
\usage{
save_datar(..., path = datar_path(), create = TRUE)
}
\arguments{
\item{...}{Objects to save}

\item{path}{Path to \code{datar} directory in which objects will be \code{save}d, defaults to \verb{/path/to/project/data/r}}
}
\description{
Save objects to datar
}
