% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_to_data_repository}
\alias{save_to_data_repository}
\title{Save objects to a data repository}
\usage{
save_to_data_repository(..., path = data_repository_path())
}
\arguments{
\item{...}{Object(s) to save}

\item{path}{Path to 'data' subdirectory of a data repository in which objects will be \code{save}d, defaults to \code{datarepo:::data_repository_path()}}
}
\value{
Invisibly returns paths to saved object(s).
}
\description{
Saves an \code{rda} in the 'data' directory of a data repository.
}
