% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\docType{data}
\name{saved_object_digests}
\alias{saved_object_digests}
\title{The list of object digests}
\format{
An object of class \code{list} of length 0.
}
\usage{
saved_object_digests
}
\description{
The list of object digests
}
\keyword{datasets}
