% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_summary.R
\name{summary_df}
\alias{summary_df}
\title{Data Frame Summary}
\usage{
summary_df(data, col, label)
}
\arguments{
\item{data}{a data frame.}

\item{col}{a name of column containing the factor interest.}

\item{label}{a string label for factor.}
}
\description{
Print summary statistics to console.
}
\author{
Nathan Lucas
}
