% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Create the data repository}
\usage{
init(path = data_repository_path())
}
\arguments{
\item{path}{Path to data repository}
}
\value{
Invisibly returns \code{path}
}
\description{
Create the data repository
}
