% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_to_data_repository}
\alias{save_to_data_repository}
\alias{save}
\title{Save objects to a data repository}
\usage{
save_to_data_repository(..., path = data_repository_path())

save(...)
}
\arguments{
\item{...}{Arguments passed to \code{base::save} or \code{save_to_data_repository()} including objects to save}

\item{path}{Path to 'data' subdirectory of a data repository in which objects will be \code{save}d, defaults to \code{datarepo:::data_repository_path()}}
}
\value{
Invisibly returns paths to saved object(s).
}
\description{
Saves an \code{rda} in the 'data' directory of a data repository.

Uses either the \code{base} or \code{datarepository} versions of \code{save} depending on arguments. If the \code{file} argument is used, it is assumed that \code{base::save} was intended, otherwise the \code{datarepository} version is used.
}
\section{Functions}{
\itemize{
\item \code{save_to_data_repository}: Save to data repository

\item \code{save}: Save data using \code{base::save}
}}

