% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save}
\alias{save}
\title{Save objects to a data repository}
\usage{
save(...)
}
\description{
Uses either the \code{base} or \code{datarepository} versions of \code{save} depending on arguments. If the \code{file} argument is used, it is assumed that \code{base::save} was intended, otherwise the \code{datarepository} version is used.
}
